Major update/cleanup to Formsy API; React v15+; No more mixins (#470)
* Up node to 8.2.1; only allow React v15+

* Update HOC to ES6 component; kill Mixin and decorator

* Update all tests; ES6ify everything!

* Update examples to ES6 classes

* Update API docs

* Fix broken/commented test case

* Move peer dep to regular dep since peer is deprecated
