+++ .nvmrc
@@ -1,0 +1,1 @@
+8.2.1
