+++ .travis.yml
@@ -1,4 +1,4 @@
 language: node_js
 sudo: false
 node_js:
-  - '4.1'
+  - '8.2.1'
