--- bower.json
@@ -1,25 +1,0 @@
-{
-  "name": "formsy-react",
-  "version": "0.18.0",
-  "description": "A form input builder and validator for React JS",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/christianalfoni/formsy-react.git"
-  },
-  "main": "src/main.js",
-  "license": "MIT",
-  "ignore": [
-    "build/",
-    "Gulpfile.js"
-  ],
-  "dependencies": {
-    "react": "^0.14.7 || ^15.0.0"
-  },
-  "keywords": [
-    "react",
-    "form",
-    "forms",
-    "validation",
-    "react-component"
-  ]
-}
