+++ examples/components/Input.js
@@ -1,29 +1,25 @@
 import React from 'react';
 import Formsy from 'formsy-react';
 
-const MyInput = React.createClass({
-
-  // Add the Formsy Mixin
-  mixins: [Formsy.Mixin],
-
+class MyInput extends React.Component {
   // setValue() will set the value of the component, which in
   // turn will validate it and the rest of the form
-  changeValue(event) {
-    this.setValue(event.currentTarget[this.props.type === 'checkbox' ? 'checked' : 'value']);
-  },
-  render() {
+  changeValue = (event) => {
+    this.props.setValue(event.currentTarget[this.props.type === 'checkbox' ? 'checked' : 'value']);
+  }
 
+  render() {
     // Set a specific className based on the validation
     // state of this component. showRequired() is true
     // when the value is empty and the required prop is
     // passed to the input. showError() is true when the
     // value typed is invalid
     const className = 'form-group' + (this.props.className || ' ') +
-      (this.showRequired() ? 'required' : this.showError() ? 'error' : '');
+      (this.props.showRequired() ? 'required' : this.props.showError() ? 'error' : '');
 
     // An error message is returned ONLY if the component is invalid
     // or the server has returned an error message
-    const errorMessage = this.getErrorMessage();
+    const errorMessage = this.props.getErrorMessage();
 
     return (
       <div className={className}>
@@ -32,13 +28,13 @@ const MyInput = React.createClass({
           type={this.props.type || 'text'}
           name={this.props.name}
           onChange={this.changeValue}
-          value={this.getValue()}
-          checked={this.props.type === 'checkbox' && this.getValue() ? 'checked' : null}
+          value={this.props.getValue()}
+          checked={this.props.type === 'checkbox' && this.props.getValue() ? 'checked' : null}
         />
         <span className='validation-error'>{errorMessage}</span>
       </div>
     );
   }
-});
+}
 
-export default MyInput;
+export default Formsy.Wrapper(MyInput);
