+++ package.json
@@ -6,7 +6,7 @@
     "type": "git",
     "url": "https://github.com/christianalfoni/formsy-react.git"
   },
-  "main": "lib/main.js",
+  "main": "lib/index.js",
   "scripts": {
     "build": "NODE_ENV=production webpack -p --config webpack.production.config.js",
     "test": "babel-node testrunner",
@@ -23,7 +23,9 @@
     "react-component"
   ],
   "dependencies": {
-    "form-data-to-object": "^0.2.0"
+    "form-data-to-object": "^0.2.0",
+    "hoist-non-react-statics": "^2.2.1",
+    "react": "^15.0.0"
   },
   "devDependencies": {
     "babel-cli": "^6.6.5",
@@ -42,8 +44,5 @@
     "sinon": "^1.17.3",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
-  },
-  "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
   }
 }
