+++ tests/Formsy-spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 import Formsy from './..';
 import TestInput from './utils/TestInput';
@@ -9,10 +9,9 @@ import immediate from './utils/immediate';
 import sinon from 'sinon';
 
 export default {
-
   'Setting up a form': {
     'should expose the users DOM node through an innerRef prop': function (test) {
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form>
@@ -20,7 +19,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = form.name;
@@ -50,7 +49,7 @@ export default {
     'should allow for null/undefined children': function (test) {
 
       let model = null;
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
@@ -61,7 +60,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       immediate(() => {
@@ -77,17 +76,17 @@ export default {
       const inputs = [];
       let forceUpdate = null;
       let model = null;
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
               {inputs}
             </Formsy.Form>);
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait before adding the input
@@ -113,17 +112,17 @@ export default {
       const inputs = [];
       let forceUpdate = null;
       let model = null;
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
               {inputs}
             </Formsy.Form>);
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait before adding the input
@@ -151,10 +150,10 @@ export default {
       let forceUpdate = null;
       let model = null;
 
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
         render() {
           const input = <TestInput name="test" value={this.props.value} />;
 
@@ -163,7 +162,7 @@ export default {
               {input}
             </Formsy.Form>);
         }
-      });
+      }
       let form = TestUtils.renderIntoDocument(<TestForm value="foo"/>);
 
       // Wait before changing the input
@@ -223,7 +222,7 @@ export default {
           super(props);
           this.state = {rule: 'ruleA'};
         }
-        changeRule() {
+        changeRule = () => {
           this.setState({
             rule: 'ruleB'
           });
@@ -256,7 +255,7 @@ export default {
           super(props);
           this.state = {showSecondInput: false};
         }
-        addInput() {
+        addInput = () => {
           this.setState({
             showSecondInput: true
           });
@@ -357,11 +356,11 @@ export default {
 
 
     const hasChanged = sinon.spy();
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return <Formsy.Form onChange={hasChanged}></Formsy.Form>;
       }
-    });
+    }
     TestUtils.renderIntoDocument(<TestForm/>);
     test.equal(hasChanged.called, false);
     test.done();
@@ -385,17 +384,15 @@ export default {
   'should trigger onChange once when new input is added to form': function (test) {
 
     const hasChanged = sinon.spy();
-    const TestForm = React.createClass({
-      getInitialState() {
-        return {
-          showInput: false
-        };
-      },
+    class TestForm extends React.Component {
+      state = {
+        showInput: false
+      }
       addInput() {
         this.setState({
           showInput: true
         })
-      },
+      }
       render() {
         return (
           <Formsy.Form onChange={hasChanged}>
@@ -407,7 +404,7 @@ export default {
             }
           </Formsy.Form>);
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     form.addInput();
@@ -422,16 +419,16 @@ export default {
 
     'should allow elements to check if the form is disabled': function (test) {
 
-      const TestForm = React.createClass({
-        getInitialState() { return { disabled: true }; },
-        enableForm() { this.setState({ disabled: false }); },
+      class TestForm extends React.Component {
+        state = { disabled: true };
+        enableForm() { this.setState({ disabled: false }); }
         render() {
           return (
             <Formsy.Form disabled={this.state.disabled}>
               <TestInput name="foo"/>
             </Formsy.Form>);
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
@@ -447,18 +444,18 @@ export default {
 
     'should be possible to pass error state of elements by changing an errors attribute': function (test) {
 
-      const TestForm = React.createClass({
-        getInitialState() { return { validationErrors: { foo: 'bar' } }; },
-        onChange(values) {
+      class TestForm extends React.Component {
+        state = { validationErrors: { foo: 'bar' } };
+        onChange = (values) => {
             this.setState(values.foo ? { validationErrors: {} } : { validationErrors: {foo: 'bar'} });
-        },
+        }
         render() {
           return (
             <Formsy.Form onChange={this.onChange} validationErrors={this.state.validationErrors}>
               <TestInput name="foo"/>
             </Formsy.Form>);
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait for update
@@ -479,14 +476,14 @@ export default {
     'should trigger an onValidSubmit when submitting a valid form': function (test) {
 
       let isCalled = sinon.spy();
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onValidSubmit={isCalled}>
               <TestInput name="foo" validations="isEmail" value="foo@bar.com"/>
             </Formsy.Form>);
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const FoundForm = TestUtils.findRenderedComponentWithType(form, TestForm);
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(FoundForm));
@@ -498,14 +495,14 @@ export default {
     'should trigger an onInvalidSubmit when submitting an invalid form': function (test) {
 
       let isCalled = sinon.spy();
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onInvalidSubmit={isCalled}>
               <TestInput name="foo" validations="isEmail" value="foo@bar"/>
             </Formsy.Form>);
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       const FoundForm = TestUtils.findRenderedComponentWithType(form, TestForm);
@@ -523,7 +520,7 @@ export default {
     'should call onSubmit correctly': function (test) {
 
       const onSubmit = sinon.spy();
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onSubmit={onSubmit}>
@@ -532,7 +529,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(form));
@@ -544,17 +541,15 @@ export default {
     'should allow dynamic changes to false': function (test) {
 
       const onSubmit = sinon.spy();
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        }
         changeValue() {
           this.setState({
             value: false
           });
-        },
+        }
         render() {
           return (
             <Formsy.Form onSubmit={onSubmit}>
@@ -563,7 +558,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       form.changeValue();
@@ -575,7 +570,7 @@ export default {
 
     'should say the form is submitted': function (test) {
 
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form>
@@ -584,7 +579,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       test.equal(input.isFormSubmitted(), false);
@@ -596,17 +591,15 @@ export default {
 
     'should be able to reset the form to its pristine state': function (test) {
 
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        }
         changeValue() {
           this.setState({
             value: false
           });
-        },
+        }
         render() {
           return (
             <Formsy.Form>
@@ -615,7 +608,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
@@ -631,17 +624,15 @@ export default {
 
     'should be able to reset the form using custom data': function (test) {
 
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        }
         changeValue() {
           this.setState({
             value: false
           });
-        },
+        }
         render() {
           return (
             <Formsy.Form>
@@ -650,7 +641,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
@@ -670,7 +661,7 @@ export default {
 
   'should be able to reset the form to empty values': function (test) {
 
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -679,7 +670,7 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const input = TestUtils.findRenderedComponentWithType(form, TestInput);
     const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
