+++ tests/Rules-equals-spec.js
@@ -1,17 +1,17 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import immediate from './utils/immediate';
 
 import Formsy from './..';
 import { InputFactory } from './utils/TestInput';
 
 const TestInput = InputFactory({
-  render() {
-    return <input value={this.getValue()} readOnly/>;
+  render: function() {
+    return <input value={this.props.getValue()} readOnly />;
   }
 });
 
-const TestForm = React.createClass({
+class TestForm extends React.Component {
   render() {
     return (
       <Formsy.Form>
@@ -19,7 +19,7 @@ const TestForm = React.createClass({
       </Formsy.Form>
     );
   }
-});
+}
 
 export default {
