+++ tests/Validation-spec.js
@@ -1,10 +1,23 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 import Formsy from './..';
-import TestInput, {InputFactory} from './utils/TestInput';
+import { InputFactory } from './utils/TestInput';
 import immediate from './utils/immediate';
 import sinon from 'sinon';
+
+class MyTest extends React.Component {
+    static defaultProps = { type: 'text' };
+
+    handleChange = (event) => {
+        this.props.setValue(event.target.value);
+    }
+
+    render() {
+      return <input type={this.props.type} value={this.props.getValue()} onChange={this.handleChange}/>;
+    }
+}
+const FormsyTest = Formsy.Wrapper(MyTest);
 
 export default {
 
@@ -12,13 +25,13 @@ export default {
 
     const form = TestUtils.renderIntoDocument(
       <Formsy.Form onSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar', bar: 'foo' })}>
-        <TestInput name="foo"/>
-        <TestInput name="bar"/>
+        <FormsyTest name="foo"/>
+        <FormsyTest name="bar"/>
       </Formsy.Form>
     );
 
     const input = TestUtils.scryRenderedDOMComponentsWithTag(form, 'INPUT')[0];
-    const inputComponents = TestUtils.scryRenderedComponentsWithType(form, TestInput);
+    const inputComponents = TestUtils.scryRenderedComponentsWithType(form, FormsyTest);
 
     form.submit();
     test.equal(inputComponents[0].isValid(), false);
@@ -37,12 +50,12 @@ export default {
 
     const form = TestUtils.renderIntoDocument(
       <Formsy.Form onSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
-        <TestInput name="foo" validations="isEmail"/>
+        <FormsyTest name="foo" validations="isEmail"/>
       </Formsy.Form>
     );
 
     const input = TestUtils.findRenderedDOMComponentWithTag(form, 'INPUT');
-    const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
+    const inputComponent = TestUtils.findRenderedComponentWithType(form, FormsyTest);
 
     form.submit();
     test.equal(inputComponent.isValid(), false);
@@ -63,7 +76,7 @@ export default {
 
     TestUtils.renderIntoDocument(
       <Formsy.Form onValid={onValid} onInvalid={onInvalid}>
-        <TestInput name="foo" value="bar" required/>
+        <FormsyTest name="foo" value="bar" required/>
       </Formsy.Form>
     );
 
@@ -80,7 +93,7 @@ export default {
 
     TestUtils.renderIntoDocument(
       <Formsy.Form onValid={onValid} onInvalid={onInvalid}>
-        <TestInput name="foo" required />
+        <FormsyTest name="foo" required />
       </Formsy.Form>
     );
 
@@ -94,8 +107,8 @@ export default {
 
     let isValid = false;
     const CustomInput = InputFactory({
-      componentDidMount() {
-        isValid = this.isValid();
+      componentDidMount: function() {
+        isValid = this.props.isValid();
       }
     });
     const form = TestUtils.renderIntoDocument(
@@ -112,20 +125,20 @@ export default {
 
   'should provide invalidate callback on onValiSubmit': function (test) {
 
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onValidSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
-            <TestInput name="foo" value="foo"/>
+            <FormsyTest name="foo" value="foo"/>
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
-    const input = TestUtils.findRenderedComponentWithType(form, TestInput);
+    const input = TestUtils.findRenderedComponentWithType(form, FormsyTest);
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), false);
     test.done();
@@ -134,19 +147,19 @@ export default {
 
   'should provide invalidate callback on onInvalidSubmit': function (test) {
 
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onInvalidSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
-            <TestInput name="foo" value="foo" validations="isEmail"/>
+            <FormsyTest name="foo" value="foo" validations="isEmail"/>
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
-    const input = TestUtils.findRenderedComponentWithType(form, TestInput);
+    const input = TestUtils.findRenderedComponentWithType(form, FormsyTest);
     TestUtils.Simulate.submit(formEl);
     test.equal(input.getErrorMessage(), 'bar');
 
@@ -156,21 +169,21 @@ export default {
 
   'should not invalidate inputs on external errors with preventExternalInvalidation prop': function (test) {
 
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form
             preventExternalInvalidation
             onSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
-            <TestInput name="foo" value="foo"/>
+            <FormsyTest name="foo" value="foo"/>
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
-    const input = TestUtils.findRenderedComponentWithType(form, TestInput);
+    const input = TestUtils.findRenderedComponentWithType(form, FormsyTest);
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), true);
     test.done();
@@ -179,19 +192,19 @@ export default {
 
   'should invalidate inputs on external errors without preventExternalInvalidation prop': function (test) {
 
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
-            <TestInput name="foo" value="foo"/>
+            <FormsyTest name="foo" value="foo"/>
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
-    const input = TestUtils.findRenderedComponentWithType(form, TestInput);
+    const input = TestUtils.findRenderedComponentWithType(form, FormsyTest);
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), false);
     test.done();
